requirements.txt
nose
jsonschema==2.3.0
PyYaml==3.11
requests
docker-py